refactor: split UserSelect into sub components
