+++ src/Molecules/UserSelect/User/index.js
@@ -1,0 +1,3 @@
+import User from './User'
+
+export default User
