+++ src/Molecules/UserSelect/UserChip/index.js
@@ -1,0 +1,3 @@
+import UserChip from './UserChip'
+
+export default UserChip
