+++ src/Molecules/UserSelect/UserChips/index.js
@@ -1,0 +1,3 @@
+import UserChips from './UserChips'
+
+export default UserChips
