+++ src/Molecules/UserSelect/UsersContext/UsersContext.jsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+
+const UsersContext = createContext({})
+
+export default UsersContext
