+++ src/Molecules/UserSelect/UsersContext/index.js
@@ -1,0 +1,3 @@
+import UsersContext from './UsersContext'
+
+export default UsersContext
