+++ src/Molecules/UserSelect/index.js
@@ -1,4 +1,8 @@
-import UserSelect, { User } from './UserSelect'
+import UserSelect from './UserSelect'
 
-export { User }
+export { default as User } from './User'
+export { default as UserChip } from './UserChip'
+export { default as UserChips } from './UserChips'
+export { default as UsersContext } from './UsersContext'
+
 export default UserSelect
